requirements.txt

PyPDF2
transformers
torch
opencv-python
pytesseract
Pillow
pdf2image